Build system mostly finished, need refactoring.
The build system has been mostly rewritten at this point, minus a few
targets. The toolchain is integrated into the build process and the
build system now supports multiple targets.

The code for the hard disk bootsector is refactored, however the rest of
the project (floppy disk bootsector, Gloss main module, and the little
bit of the Slick kernel that we have) still needs refactoring to fit the
new build process.